rewrite the project and migrate to uv
load titles from file

load titles from yaml instead of json

add option to specify interval in the config file

make the script follow XDG path specifications for the state-, config- and log files

make the script print the config-, state- and log file paths on startup

add option for a oneshot check

Add a commandline flag '--once'. If this is passed to the script, it
will run the checks once and exit, instead of running a loop.

add desktop notification sending functionality

When change is detected in the number of episodes and when a title is
first checked a desktop notification will be sent.

restructure repo

migrate the project from using hatch as a package manager to uv

update readme

rename workflow

update workflow

activate workflow

update workflow

update readme

update version